color detection
Color detection is necessary to recognize objects, it is also used as a tool in various image editing and drawing apps. In this color detection Python project, we are going to build an application through which you can automatically get the name of the color by clicking on them. So for this, we will have a data file that contains the color name and its values. Then we will calculate the distance from each color and find the shortest one.