Gender Detection
To build a gender and age detector that can approximately guess the gender and age of the person (face) in a picture or through webcam. In this Python Project, I have used Deep Learning to accurately identify the gender and age of a person from a single image of a face.